:recycle: getDeterms
